fix(config): expose TUI validation through config API
Plan: .omo/plans/omo-tui-sidebar-panel.md
